refactor: colocate home components
