+++ src/app/(platform)/(home)/_components/EventCard.tsx
@@ -6,13 +6,13 @@ import { useAppKitAccount } from '@reown/appkit/react'
 import { useQueryClient } from '@tanstack/react-query'
 import { use, useMemo, useState } from 'react'
 import { useSignTypedData } from 'wagmi'
+import EventCardFooter from '@/app/(platform)/(home)/_components/EventCardFooter'
+import EventCardHeader from '@/app/(platform)/(home)/_components/EventCardHeader'
+import EventCardMarketsList from '@/app/(platform)/(home)/_components/EventCardMarketsList'
+import EventCardSingleMarketActions from '@/app/(platform)/(home)/_components/EventCardSingleMarketActions'
+import EventCardTradePanel from '@/app/(platform)/(home)/_components/EventCardTradePanel'
+import { OpenCardContext } from '@/app/(platform)/(home)/_components/EventOpenCardProvider'
 import { handleOrderCancelledFeedback, handleOrderErrorFeedback, handleOrderSuccessFeedback, handleValidationError, notifyWalletApprovalPrompt } from '@/app/(platform)/event/[slug]/_components/feedback'
-import EventCardFooter from '@/components/EventCardFooter'
-import EventCardHeader from '@/components/EventCardHeader'
-import EventCardMarketsList from '@/components/EventCardMarketsList'
-import EventCardSingleMarketActions from '@/components/EventCardSingleMarketActions'
-import EventCardTradePanel from '@/components/EventCardTradePanel'
-import { OpenCardContext } from '@/components/EventOpenCardContext'
 import { Card, CardContent } from '@/components/ui/card'
 import { useAffiliateOrderMetadata } from '@/hooks/useAffiliateOrderMetadata'
 import { useAppKit } from '@/hooks/useAppKit'
