+++ src/app/(platform)/(home)/_components/EventCardFooter.tsx
