+++ src/app/(platform)/(home)/_components/EventCardHeader.tsx
