+++ src/app/(platform)/(home)/_components/EventCardMarketsList.tsx
