+++ src/app/(platform)/(home)/_components/EventCardSkeleton.tsx
