+++ src/app/(platform)/(home)/_components/EventCardTradePanel.tsx
