+++ src/app/(platform)/(home)/_components/EventOpenCardProvider.tsx
