+++ src/app/(platform)/(home)/_components/EventsGrid.tsx
@@ -5,13 +5,13 @@ import type { Event } from '@/types'
 import { useInfiniteQuery } from '@tanstack/react-query'
 import { useWindowVirtualizer } from '@tanstack/react-virtual'
 import { useEffect, useMemo, useRef, useState } from 'react'
+import EventCard from '@/app/(platform)/(home)/_components/EventCard'
+import EventCardSkeleton from '@/app/(platform)/(home)/_components/EventCardSkeleton'
+import EventsGridSkeleton from '@/app/(platform)/(home)/_components/EventsGridSkeleton'
 import EventsEmptyState from '@/app/(platform)/event/[slug]/_components/EventsEmptyState'
 import { useEventLastTrades } from '@/app/(platform)/event/[slug]/_hooks/useEventLastTrades'
 import { useEventMarketQuotes } from '@/app/(platform)/event/[slug]/_hooks/useEventMidPrices'
 import { buildMarketTargets } from '@/app/(platform)/event/[slug]/_hooks/useEventPriceHistory'
-import EventCard from '@/components/EventCard'
-import EventCardSkeleton from '@/components/EventCardSkeleton'
-import EventsGridSkeleton from '@/components/EventsGridSkeleton'
 import { useColumns } from '@/hooks/useColumns'
 import { resolveDisplayPrice } from '@/lib/market-chance'
 import { cn } from '@/lib/utils'
