+++ src/app/(platform)/(home)/_components/FilterToolbar.tsx
@@ -5,7 +5,7 @@ import type { FilterState } from '@/providers/FilterProvider'
 import { useAppKitAccount } from '@reown/appkit/react'
 import { BookmarkIcon, ClockIcon, DropletIcon, FlameIcon, HandFistIcon, Settings2Icon, SparklesIcon, TrendingUpIcon } from 'lucide-react'
 import { useCallback, useEffect, useMemo, useState } from 'react'
-import FilterToolbarSearchInput from '@/components/FilterToolbarSearchInput'
+import FilterToolbarSearchInput from '@/app/(platform)/(home)/_components/FilterToolbarSearchInput'
 import { Button } from '@/components/ui/button'
 import { Checkbox } from '@/components/ui/checkbox'
 import { Label } from '@/components/ui/label'
