+++ src/app/(platform)/(home)/_components/HomeClient.tsx
@@ -2,9 +2,9 @@
 
 import type { Event } from '@/types'
 import { useEffect, useRef } from 'react'
-import { OpenCardProvider } from '@/components/EventOpenCardContext'
-import EventsGrid from '@/components/EventsGrid'
-import FilterToolbar from '@/components/FilterToolbar'
+import { OpenCardProvider } from '@/app/(platform)/(home)/_components/EventOpenCardProvider'
+import EventsGrid from '@/app/(platform)/(home)/_components/EventsGrid'
+import FilterToolbar from '@/app/(platform)/(home)/_components/FilterToolbar'
 import { useFilters } from '@/providers/FilterProvider'
 
 interface HomeClientProps {
